Fix makerelease script
The makerelease script was lacking error handling.
This is added in order to detect failures on
intermediate steps.

+++ bin/makeRelease.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-# @todo add more error handling
 
 PREVIOUS_SF_ENV=${SYMFONY_ENV}
 PREVIOUS_EB_ENV=${ENGINEBLOCK_ENV}
@@ -36,6 +35,13 @@ fi
 PROJECT_DIR_NAME=${PROJECT_NAME}-${TAG//\//_} &&
 PROJECT_DIR=${RELEASE_DIR}/${PROJECT_DIR_NAME} &&
 
+# Check requirements
+command -v php72 >/dev/null 2>&1 || { echo >&2 "Missing PHP 7.2. Aborting"; exit 1; }
+command -v composer >/dev/null 2>&1 || { echo >&2 "Missing Composer.  Aborting."; exit 1; }
+command -v npm >/dev/null 2>&1 || { echo >&2 "Misisng NPM.  Aborting."; exit 1; }
+command -v git >/dev/null 2>&1 || { echo >&2 "Misisng Git.  Aborting."; exit 1; }
+
+# Prepare environment
 echo "Preparing environment" &&
 mkdir -p ${RELEASE_DIR} &&
 rm -rf ${PROJECT_DIR} &&
@@ -46,25 +52,53 @@ git clone https://github.com/${GITHUB_USER}/${PROJECT_NAME}.git ${PROJECT_DIR_NA
 
 echo "Checking out ${TAG}" &&
 cd ${PROJECT_DIR} &&
-git checkout ${TAG} &&
+git checkout ${TAG}
 
+if [ $? -eq 0 ]; then
+    echo "Project prepared"
+else
+    echo "Initialization failed"
+    exit 1
+fi
+
+
+# Install composer dependencies
 echo "Running Composer Install" &&
-php72 ./bin/composer.phar install -n --no-dev --prefer-dist -o &&
+php72 $(which composer) install -n --no-dev --prefer-dist -o
 
+if [ $? -eq 0 ]; then
+    echo "Composer install ran"
+else
+    echo "Unable to run compopser install"
+    exit 1
+fi
+
+
+# Build NPM frontend assets
 # --unsafe-perm because we do branch install as root.
 # can be removed when we stop doing that
-echo "Build assets" &&
+echo "Build assets"
 cd ${PROJECT_DIR}/theme &&
 npm ci --unsafe-perm &&
-npm run release &&
+npm run release
 
+if [ $? -eq 0 ]; then
+    echo "Assets build"
+else
+    echo "Unable to build assets"
+    exit 1
+fi
+
+
+# Tag release and remove unwanted files
 echo "Tagging the release in RELEASE file" &&
 COMMITHASH=`git rev-parse HEAD` &&
 echo "Tag: ${TAG}" > ${PROJECT_DIR}/RELEASE &&
 echo "Commit: ${COMMITHASH}" >> ${PROJECT_DIR}/RELEASE &&
 
-echo "Updating asset_version in config"
-sed -i s,#ASSET_VERSION#,${TAG},g ${PROJECT_DIR}/app/config/config.yml
+echo "Updating asset_version in config" &&
+sed -i s,#ASSET_VERSION#,${TAG},g ${PROJECT_DIR}/app/config/config.yml &&
+
 
 echo "Cleaning build of dev files" &&
 rm -rf ${PROJECT_DIR}/.idea &&
@@ -80,13 +114,30 @@ rm -rf ${PROJECT_DIR}/tests &&
 rm -rf ${PROJECT_DIR}/ci &&
 rm -rf ${PROJECT_DIR}/.travis.yml &&
 rm -rf ${PROJECT_DIR}/theme/node_modules &&
-rm -rf ${PROJECT_DIR}/theme/.sass-cache &&
+rm -rf ${PROJECT_DIR}/theme/.sass-cache
 
+if [ $? -eq 0 ]; then
+    echo "Release buildd"
+else
+    echo "Failed to build release"
+    exit 1
+fi
+
+
+# Create tarball
 echo "Create tarball" &&
 cd ${RELEASE_DIR} &&
 tar -czf ${PROJECT_DIR_NAME}.tar.gz ${PROJECT_DIR_NAME}
 
+if [ $? -eq 0 ]; then
+    echo "Tarball build"
+else
+    echo "Unable to build tarball"
+    exit 1
+fi
 
+
+# Create checksum
 echo "Create checksum file" &&
 cd ${RELEASE_DIR} &&
 if hash sha1sum 2>/dev/null; then
@@ -95,6 +146,15 @@ else
     shasum ${PROJECT_DIR_NAME}.tar.gz > ${PROJECT_DIR_NAME}.sha
 fi
 
+if [ $? -eq 0 ]; then
+    echo "Checksum created"
+else
+    echo "Unable to create checksum"
+    exit 1
+fi
+
+
+# Sign with GPG key
 if [ -n "$2" ]
 then
 	if [ "$2" == "sign" ]
@@ -102,6 +162,14 @@ then
 	    echo "Signing build"
 		cd ${RELEASE_DIR}
 		gpg -o ${PROJECT_DIR_NAME}.sha.gpg  --clearsign ${PROJECT_DIR_NAME}.sha
+
+        if [ $? -eq 0 ]; then
+            echo "Signed"
+        else
+            echo "Unable to sign tarball"
+            exit 1
+        fi
+
 	fi
 fi
